finished building indexing for projects

+++ app/src/main/java/com/door43/translationstudio/tasks/IndexProjectsTask.java
@@ -1,5 +1,6 @@
 package com.door43.translationstudio.tasks;
 
+import com.door43.translationstudio.R;
 import com.door43.translationstudio.projects.Chapter;
 import com.door43.translationstudio.projects.Frame;
 import com.door43.translationstudio.projects.Model;
@@ -15,6 +16,7 @@ import org.json.JSONException;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 
 /**
  * This task indexes a project so we don't have to load the entire thing into memory.
@@ -37,34 +39,43 @@ public class IndexProjectsTask extends ManagedTask {
         for(Project proj:mProjects) {
             pIndex ++;
             if(interrupted()) return;
-            // TRICKY: we create a new project so we don't override anything the user is working on.
-            Project p = proj.softClone();
-            File projectDir = new File(indexDir, p.getId());
+            File projectDir = new File(indexDir, proj.getId());
+            File projectReadyFile = new File(projectDir, "ready.index");
+            if(projectReadyFile.exists()) {
+                continue;
+            }
             projectDir.mkdirs();
 
             // index project
             File pInfo = new File(projectDir, "data.json");
-            try {
-                FileUtils.write(pInfo, p.serialize().toString());
-            } catch (Exception e) {
-                Logger.e(this.getClass().getName(), "Failed to index project info. Project: " + p.getId(), e);
-                continue;
+            if(!pInfo.exists()) {
+                try {
+                    FileUtils.write(pInfo, proj.serialize().toString());
+                } catch (Exception e) {
+                    Logger.e(this.getClass().getName(), "Failed to index project info. Project: " + proj.getId(), e);
+                    continue;
+                }
             }
 
             int lIndex = -1;
-            for(SourceLanguage l:p.getSourceLanguages()) {
+            for(SourceLanguage l:proj.getSourceLanguages()) {
                 lIndex ++;
                 if(interrupted()) return;
                 File languageDir = new File(projectDir, l.getId());
                 languageDir.mkdirs();
 
+                // TRICKY: we create a new project so we don't override anything the user is working on.
+                Project p = proj.softClone();
+
                 // index language
                 File lInfo = new File(languageDir, "data.json");
-                try {
-                    FileUtils.write(lInfo, p.serializeSourceLanguage(l).toString());
-                } catch (Exception e) {
-                    Logger.e(this.getClass().getName(), "Failed to index language info. Project: " + p.getId() + " Language: " + l.getId(), e);
-                    continue;
+                if(!lInfo.exists()) {
+                    try {
+                        FileUtils.write(lInfo, p.serializeSourceLanguage(l).toString());
+                    } catch (Exception e) {
+                        Logger.e(this.getClass().getName(), "Failed to index language info. Project: " + p.getId() + " Language: " + l.getId(), e);
+                        continue;
+                    }
                 }
 
                 int rIndex = -1;
@@ -76,15 +87,17 @@ public class IndexProjectsTask extends ManagedTask {
 
                     // index resource
                     File rInfo = new File(resourceDir, "data.json");
-                    try {
-                        FileUtils.write(rInfo, r.serialize().toString());
-                    } catch (Exception e) {
-                        Logger.e(this.getClass().getName(), "Failed to index resource info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId(), e);
-                        continue;
+                    if(!rInfo.exists()) {
+                        try {
+                            FileUtils.write(rInfo, r.serialize().toString());
+                        } catch (Exception e) {
+                            Logger.e(this.getClass().getName(), "Failed to index resource info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId(), e);
+                            continue;
+                        }
                     }
 
                     // load the source
-                    AppContext.projectManager().fetchProjectSource(p, false);
+                    AppContext.projectManager().fetchProjectSource(p, l, r, false);
                     File sourceDir = new File(resourceDir, "source");
                     File notesDir = new File(resourceDir, "notes");
 
@@ -97,16 +110,15 @@ public class IndexProjectsTask extends ManagedTask {
                         File notesChaptersDir = new File(notesDir, c.getId());
                         notesChaptersDir.mkdirs();
 
-                        // NOTE: we publish the progress here so we don't flood the ui with updates
-                        publishProgress((pIndex + (lIndex + (rIndex + (cIndex + 1) / (double) p.getChapters().length) / (double) l.getResources().length) / (double) p.getSourceLanguages().size()) / (double) mProjects.length, "Indexing projects...");
-
                         // index chapter
                         File cInfo = new File(sourceChaptersDir, "data.json");
-                        try {
-                            FileUtils.write(cInfo, c.serialize().toString());
-                        } catch (Exception e) {
-                            Logger.e(this.getClass().getName(), "Failed to index chapter info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId() + " Chapter: " + c.getId(), e);
-                            continue;
+                        if(!cInfo.exists()) {
+                            try {
+                                FileUtils.write(cInfo, c.serialize().toString());
+                            } catch (Exception e) {
+                                Logger.e(this.getClass().getName(), "Failed to index chapter info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId() + " Chapter: " + c.getId(), e);
+                                continue;
+                            }
                         }
 
                         // index frames
@@ -114,26 +126,45 @@ public class IndexProjectsTask extends ManagedTask {
                         for(Model m:c.getFrames()) {
                             fIndex ++;
                             if(interrupted()) return;
-                            sleep(100);
-                            if(interrupted()) return;
 
                             Frame f = (Frame)m;
                             File sourceFrameInfo = new File(sourceChaptersDir, f.getId() + ".json");
                             File notesFrameInfo = new File(notesChaptersDir, f.getId() + ".json");
 
-                            try {
-                                FileUtils.write(sourceFrameInfo, f.serialize().toString());
-                            } catch (Exception e) {
-                                Logger.e(this.getClass().getName(), "Failed to index source frame info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId() + " Chapter: " + c.getId() + " Frame: " + f.getId(), e);
+                            publishProgress((pIndex + (lIndex + (rIndex + (cIndex + (fIndex + 1) / (double) c.getFrames().length) / (double) p.getChapters().length) / (double) l.getResources().length) / (double) p.getSourceLanguages().size()) / (double) mProjects.length, AppContext.context().getResources().getString(R.string.indexing_projects) + " " + p.getId() + "." + l.getId() + "." + c.getId());
+
+
+                            if(!sourceFrameInfo.exists()) {
+                                try {
+                                    FileUtils.write(sourceFrameInfo, f.serialize().toString());
+                                } catch (Exception e) {
+                                    Logger.e(this.getClass().getName(), "Failed to index source frame info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId() + " Chapter: " + c.getId() + " Frame: " + f.getId(), e);
+                                }
                             }
 
-                            try {
-                                FileUtils.write(notesFrameInfo, f.serializeTranslationNote().toString());
-                            } catch (Exception e) {
-                                Logger.e(this.getClass().getName(), "Failed to index notes frame info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId() + " Chapter: " + c.getId() + " Frame: " + f.getId(), e);
+                            if(!notesFrameInfo.exists()) {
+                                try {
+                                    if(f.getTranslationNotes() != null) {
+                                        FileUtils.write(notesFrameInfo, f.serializeTranslationNote().toString());
+                                    }
+                                } catch (Exception e) {
+                                    Logger.e(this.getClass().getName(), "Failed to index notes frame info. Project: " + p.getId() + " Language: " + l.getId() + " Resource: " + r.getId() + " Chapter: " + c.getId() + " Frame: " + f.getId(), e);
+                                }
                             }
                         }
                     }
+
+                    // empty the source
+                    p.flush();
+                }
+            }
+
+            // mark the index as complete
+            if(!interrupted()) {
+                try {
+                    FileUtils.write(projectReadyFile, proj.getDateModified() + "");
+                } catch (IOException e) {
+                    Logger.e(this.getClass().getName(), "Failed to create the ready.index file", e);
                 }
             }
         }
